requirements.txt
beautifulsoup4==4.14.2
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
colorama==0.4.6
curl_cffi==0.13.0
frozendict==2.4.6
greenlet==3.2.4
html5lib==1.1
idna==3.11
lxml==6.0.2
multitasking==0.0.12
numpy==2.0.2
pandas==2.2.2
peewee==3.18.3
platformdirs==4.5.0
protobuf==6.33.0
psycopg2-binary==2.9.9
pycparser==2.23
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.2
requests==2.32.3
six==1.17.0
soupsieve==2.8
SQLAlchemy==2.0.31
tqdm==4.67.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
webencodings==0.5.1
websockets==15.0.1
yfinance==0.2.66
